scarabintheloop/requirements.txt
python-slugify 
